requirements.txt
django==1.11.18
djangorestframework==3.8.2
django-environ==0.4.0
psycopg2==2.7.3.2
